requirements.txt
python-telegram-bot==21.6
gspread==6.1.4
google-auth==2.36.0
httpx==0.28.1
python-dotenv==1.0.1
